Fold the paste undo-cost test into the existing undo suite
A new frontend test file that mounts the docs editor costs the suite
another full @wafflebase/docs module graph, transformed in parallel
with every other file. That was enough to reliably time out the
unrelated 5s 'TextEditSection module imports without error' smoke
test — the suite passed with the file removed and failed with it
present, and shrinking the work inside the test changed nothing.

editor-undo-selection.test.ts already imports exactly the same
modules and is already the home for editor-level undo regressions
driven through a real YorkieDocStore, so the test moves there and
shares its canvas shim.
